fixed youtube embeds
We tried using the lazy="loading" attribute but it still makes a lot of
requests to YT even if the user hasn't scrolled to the YT video in his
viewport, not to mention the number of requests when he does scroll to
the video even if he doesn't play it.

We tried using the labnol lite YT embed but that didn't work as expected
so we went with paulirish/lite-youtube-embed which works fine. Thanks to
paulirish for making this.